Warning free.

+++ swoole_mysql_coro.cc
@@ -1314,7 +1314,7 @@ static PHP_METHOD(swoole_mysql_coro_statement, nextResult)
 
     mysql_client *client = stmt->client;
 
-    if (stmt->buffer && stmt->buffer->offset < stmt->buffer->length)
+    if (stmt->buffer && (size_t) stmt->buffer->offset < stmt->buffer->length)
     {
         client->cmd = SW_MYSQL_COM_STMT_EXECUTE;
         client->state = SW_MYSQL_STATE_READ_START;
@@ -1731,7 +1731,7 @@ static int swoole_mysql_coro_onHandShake(mysql_client *client TSRMLS_DC)
         }
 
         // may be more packages
-        if (buffer->offset < buffer->length)
+        if ((size_t) buffer->offset < buffer->length)
         {
             goto _again;
         }
@@ -1842,7 +1842,7 @@ static int swoole_mysql_coro_onRead(swReactor *reactor, swEvent *event)
                 case SW_CLOSE:
                     goto close_fd;
                 case SW_WAIT:
-                    if (client->check_offset == buffer->length)
+                    if ((size_t) client->check_offset == buffer->length)
                     {
                         return SW_OK;
                     }
